requirements.txt
Django>=3.1.5,<3.2.0
djangorestframework>=3.12.2,<3.13
psycopg2>=2.8.6,<2.9.0
psycopg2-binary>=2.8.6,<2.9.0